Separate folders for map data
Data files are in separate folders depending on the map and only
downloaded if they do not exist yet.
Folders are also created if a map is used for the first time.
Special case for the maps integrated in ASCN global map which all use
data for all the genes of the global map.

Refactor the code to make it simpler to maintain.